Refactor, decouple autocomplete components
