+++ tests/unit/components/paper-autocomplete-test.js
@@ -1,4 +1,3 @@
-import Ember from 'ember';
 import { moduleForComponent, test } from 'ember-qunit';
 
 moduleForComponent('paper-autocomplete', 'Unit | Component | paper autocomplete', {
@@ -41,42 +40,5 @@ test('it propagates placeholder to input box', function(assert) {
   this.render();
   assert.equal(component.$().find('input').attr('placeholder'), 'Testing', 'Sets correct placeholder on input box.');
 });
-
-
-test('have ulContainer correctly registered', function(assert) {
-  assert.expect(2);
-
-  // Creates the component instance
-  var component = this.subject();
-
-  this.render();
-
-  assert.equal(component.get('ulContainer')._state, 'inDOM', 'ulContainer is registered in DOM.');
-
-  assert.equal(component.get('ulContainer').$().parent().prop('tagName'), 'BODY', 'ulContainer is placed inside the body tag.');
-
-});
-
-test('can search for Array of strings', function(assert) {
-  assert.expect(2);
-
-  // Creates the component instance
-  var component = this.subject({
-    source: Ember.A(['One','Two','Three']),
-  });
-  this.render();
-
-
-  Ember.run.next(function () {
-    component.set('searchText', 'On');
-
-    Ember.run.next(function () {
-      assert.equal(component.get('suggestions').length, 1, 'Has one suggestion for given criteria');
-      assert.equal(component.get('ulContainer').$().is(':visible'), true, 'list of suggestions is visible');
-    });
-  });
-
-});
-
